+++ includes/graphic.h
@@ -1,0 +1,14 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   graphic.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: nrontey <nrontey@student.42angouleme.fr    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/10/14 16:01:43 by nrontey           #+#    #+#             */
+/*   Updated: 2024/10/14 16:20:27 by nrontey          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+// mlx_make.c
+int		make_the_window(t_data *data);
